src/requirements.txt
# Azure dependencies
opencensus-ext-azure==1.1.15
opencensus-context==0.1.3

# Logging and utilities
python-dotenv==1.1.1

# For potential web interface (optional)
flask==3.1.1

# Development and testing
pytest==8.4.1
pytest-mock==3.14.1

# Code quality
black==25.1.0
flake8==7.3.0